Rename DmsDocumentsServiceProvider to DocumentsServiceProvider, register also the newly added FileService
